meu-agente-app/apps/services/user-settings-service/requirements.txt
# requirements.txt - Adicione essas dependências ao seu arquivo existente

# Dependências existentes
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-ai==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
openai==1.3.8

# Novas dependências para PostgreSQL
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Dependências adicionais úteis
python-jose[cryptography]==3.3.0  # Para JWT futuramente
passlib[bcrypt]==1.7.4             # Para hash de senhas futuramente
redis==5.0.1                       # Para cache
celery==5.3.4                      # Para tarefas assíncronas

# Para desenvolvimento e testes
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # Para testar APIs